requirements.txt
# Core dependencies
numpy==1.24.3
torch==2.1.2
torchvision==0.16.2

# Image processing
Pillow==10.1.0
scikit-image==0.22.0
matplotlib==3.8.2

# Machine learning
scikit-learn==1.3.2

# CLIP and utilities
ftfy==6.1.3
regex==2023.10.3
tqdm==4.66.1
git+https://github.com/openai/CLIP.git

# COCO API
pycocotools==2.0.7

# Optional but recommended
ipython==8.12.0  # For development
jupyter==1.0.0   # For notebook experiments
pandas==2.1.4    # For data analysis